Move xarray functionality to a separate module
