+++ dask_groupby/__init__.py
@@ -3,7 +3,7 @@
 """Top-level module for dask_groupby ."""
 from pkg_resources import DistributionNotFound, get_distribution
 
-from .core import groupby_reduce, xarray_groupby_reduce, xarray_reduce  # noqa
+from .core import groupby_reduce  # noqa
 
 try:
     __version__ = get_distribution(__name__).version
