+++ tests/__init__.py
@@ -1,4 +1,7 @@
 import dask
+import dask.array as da
+import numpy as np
+import xarray as xr
 
 
 class CountingScheduler:
@@ -23,3 +26,25 @@ def raise_if_dask_computes(max_computes=0):
     # return a dummy context manager so that this can be used for non-dask objects
     scheduler = CountingScheduler(max_computes)
     return dask.config.set(scheduler=scheduler)
+
+
+def assert_equal(a, b):
+    __tracebackhide__ = True
+
+    if isinstance(a, list):
+        a = np.array(a)
+    if isinstance(b, list):
+        b = np.array(b)
+    if isinstance(a, da.Array) or isinstance(b, da.Array):
+        # does some validation of the dask graph
+        try:
+            da.utils.assert_eq(a, b)
+        except AssertionError:
+            # dask doesn't consider nans in the same place to be equal
+            # from xarray.core.duck_array_ops.array_equiv
+            flag_array = (a == b) | (np.isnan(a) & np.isnan(b))
+            assert bool(flag_array.all())
+    elif isinstance(a, xr.DataArray) | isinstance(b, xr.DataArray):
+        xr.testing.assert_identical(a, b)
+    else:
+        np.testing.assert_equal(a, b)
