+++ tests/test_core.py
@@ -2,13 +2,12 @@ import dask
 import dask.array as da
 import numpy as np
 import pytest
-import xarray as xr
 from dask.array import from_array
 from numpy_groupies.aggregate_numpy import aggregate
 
-from dask_groupby.core import groupby_reduce, reindex_, xarray_groupby_reduce, xarray_reduce
+from dask_groupby.core import groupby_reduce, reindex_
 
-from . import raise_if_dask_computes
+from . import assert_equal, raise_if_dask_computes
 
 labels = np.array([0, 0, 2, 2, 2, 1, 1, 2, 2, 1, 1, 0])
 nan_labels = labels.astype(float)  # copy
@@ -17,30 +16,9 @@ labels2d = np.array([labels[:5], np.flip(labels[:5])])
 
 dask.config.set(scheduler="sync")
 
-
-def assert_equal(a, b):
-    __tracebackhide__ = True
-
-    if isinstance(a, list):
-        a = np.array(a)
-    if isinstance(b, list):
-        b = np.array(b)
-    if isinstance(a, da.Array) or isinstance(b, da.Array):
-        # does some validation of the dask graph
-        try:
-            da.utils.assert_eq(a, b)
-        except AssertionError:
-            # dask doesn't consider nans in the same place to be equal
-            # from xarray.core.duck_array_ops.array_equiv
-            flag_array = (a == b) | (np.isnan(a) & np.isnan(b))
-            assert bool(flag_array.all())
-    elif isinstance(a, xr.DataArray) | isinstance(b, xr.DataArray):
-        xr.testing.assert_identical(a, b)
-    else:
-        np.testing.assert_equal(a, b)
-
-
 # TODO: Add max,argmax here
+
+
 @pytest.mark.parametrize("chunk, split_out", [(False, 1), (True, 1), (True, 2), (True, 3)])
 @pytest.mark.parametrize("expected_groups", [None, [0, 1, 2], np.array([0, 1, 2])])
 @pytest.mark.parametrize(
@@ -368,61 +346,3 @@ def test_bad_npg_behaviour():
         )[0]
         == -np.inf
     )
-
-
-def test_xarray_groupby_reduce():
-    arr = np.ones((4, 12))
-
-    labels = np.array(["a", "a", "c", "c", "c", "b", "b", "c", "c", "b", "b", "f"])
-    labels = np.array(labels)
-    labels2 = np.array([1, 2, 2, 1])
-
-    da = xr.DataArray(
-        arr, dims=("x", "y"), coords={"labels2": ("x", labels2), "labels": ("y", labels)}
-    ).expand_dims(z=4)
-
-    grouped = da.groupby("labels")
-    expected = grouped.mean()
-    actual = xarray_groupby_reduce(grouped, "mean")
-    assert_equal(expected, actual)
-
-    actual = xarray_groupby_reduce(da.transpose("y", ...).groupby("labels"), "mean")
-    assert_equal(expected, actual)
-
-    # TODO: fails because of stacking
-    # grouped = da.groupby("labels2")
-    # expected = grouped.mean()
-    # actual = xarray_groupby_reduce(grouped, "mean")
-    # assert_equal(expected, actual)
-
-
-def test_xarray_reduce_multiple_groupers():
-    arr = np.ones((4, 12))
-
-    labels = np.array(["a", "a", "c", "c", "c", "b", "b", "c", "c", "b", "b", "f"])
-    labels = np.array(labels)
-    labels2 = np.array([1, 2, 2, 1])
-
-    da = xr.DataArray(
-        arr, dims=("x", "y"), coords={"labels2": ("x", labels2), "labels": ("y", labels)}
-    ).expand_dims(z=4)
-
-    expected = xr.DataArray(
-        [[4, 4], [10, 10], [8, 8], [2, 2]],
-        dims=("labels", "labels2"),
-        coords={"labels": ["a", "c", "b", "f"], "labels2": [1, 2]},
-    ).expand_dims(z=4)
-
-    actual = xarray_reduce(da, da.labels, da.labels2, func="count")
-    xr.testing.assert_identical(expected, actual)
-
-    actual = xarray_reduce(da, "labels", "labels2", func="count", fill_value=0)
-    xr.testing.assert_identical(expected, actual)
-
-    with raise_if_dask_computes():
-        actual = xarray_reduce(da.chunk({"x": 2, "z": 1}), da.labels, da.labels2, func="count")
-    xr.testing.assert_identical(expected, actual)
-
-    with pytest.raises(ValueError):
-        actual = xarray_reduce(da.chunk({"x": 2, "z": 1}), "labels", "labels2", func="count")
-    # xr.testing.assert_identical(expected, actual)
